requirements.txt
streamlit==1.2.0
pandas==1.5.3
plotly-express==0.4.1